JDP240901-24 small changes, moved exeption to another pak
